chore: FloatingChatView -> Common 폴더로 이동 & RoomFloatingChatViewController 추가
