+++ Pointer_iOS/Sources/ResultScene/View/FloatChatView/FloatingChatTableViewCell.swift
@@ -7,9 +7,13 @@
 
 import UIKit
 
-class FloatChatTableViewCell: UITableViewCell {
+//MARK: - Room 채팅 테이블 뷰 셀
+class FloatingChatTableViewCell: UITableViewCell {
 
     static let identifier = "FloatChatTableViewCell"
+    
+    
+    
     
     override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
         super.init(style: style, reuseIdentifier: reuseIdentifier)
