+++ Pointer_iOS/Sources/ResultScene/View/FloatChatView/RoomFloatingChatViewController.swift
@@ -1,0 +1,57 @@
+//
+//  RoomFloatingChatViewController.swift
+//  Pointer_iOS
+//
+//  Created by 박현준 on 2023/05/12.
+//
+
+import UIKit
+
+final class RoomFloatingChatViewController: UIViewController, ScrollableViewController {
+    
+    private let tableView: SelfSizingTableView = {
+        $0.allowsSelection = false
+        $0.backgroundColor = UIColor.clear
+        $0.separatorStyle = .none
+        $0.bounces = true
+        $0.showsVerticalScrollIndicator = true
+        $0.contentInset = .zero
+        $0.indicatorStyle = .black
+        $0.estimatedRowHeight = 34.0
+        $0.register(UITableViewCell.self, forCellReuseIdentifier: "cell")
+        return $0
+    }(SelfSizingTableView(maxHeight: UIScreen.main.bounds.height * 0.62))
+    
+    var scrollView: UIScrollView {
+        tableView
+    }
+        
+    init() {
+        super.init(nibName: nil, bundle: nil)
+        setUpView()
+    }
+    required init?(coder: NSCoder) {
+        fatalError()
+    }
+    
+    private func setUpView() {
+        view.addSubview(tableView)
+        
+        tableView.snp.makeConstraints {
+            $0.edges.equalToSuperview()
+        }
+        
+        tableView.dataSource = self
+    }
+}
+
+extension RoomFloatingChatViewController: UITableViewDataSource {
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        20
+    }
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let cell = tableView.dequeueReusableCell(withIdentifier: "cell", for: indexPath)
+        cell.textLabel?.text = "cell\(indexPath.row)"
+        return cell
+    }
+}
