+++ Pointer_iOS/Sources/ResultScene/View/ResultView/ResultViewController.swift
@@ -16,7 +16,7 @@ class ResultViewController: BaseViewController {
     var viewModel = ResultViewModel()
     let disposeBag = DisposeBag()
     
-    var timeString = "2023-05-06 16:11:10"
+    var timeString = "2023-05-11 22:30:20"
     private var remainedTime = ""
     
 //MARK: - Rx
@@ -265,58 +265,9 @@ class ResultViewController: BaseViewController {
     }
     
     @objc func chatTaped() {
-        let resultChatViewController = FloatingChatViewController(contentViewController: MyViewController())
+        let resultChatViewController = FloatingChatViewController(contentViewController: RoomFloatingChatViewController())
         present(resultChatViewController, animated: true)
     }
     
     
-}
-//MARK: - ScrollableViewController: 클라이언트 코드에서 해당 프로토콜에 명시된 인터페이스에 접근 - 여기서 바꿔야 함!!
-final class MyViewController: UIViewController, ScrollableViewController {
-    
-    private let tableView: SelfSizingTableView = {
-        $0.allowsSelection = false
-        $0.backgroundColor = UIColor.clear
-        $0.separatorStyle = .none
-        $0.bounces = true
-        $0.showsVerticalScrollIndicator = true
-        $0.contentInset = .zero
-        $0.indicatorStyle = .black
-        $0.estimatedRowHeight = 34.0
-        $0.register(UITableViewCell.self, forCellReuseIdentifier: "cell")
-        return $0
-    }(SelfSizingTableView(maxHeight: UIScreen.main.bounds.height * 0.62))
-    
-    var scrollView: UIScrollView {
-        tableView
-    }
-        
-    init() {
-        super.init(nibName: nil, bundle: nil)
-        setUpView()
-    }
-    required init?(coder: NSCoder) {
-        fatalError()
-    }
-    
-    private func setUpView() {
-        view.addSubview(tableView)
-        
-        tableView.snp.makeConstraints {
-            $0.edges.equalToSuperview()
-        }
-        
-        tableView.dataSource = self
-    }
-}
-
-extension MyViewController: UITableViewDataSource {
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        20
-    }
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: "cell", for: indexPath)
-        cell.textLabel?.text = "cell\(indexPath.row)"
-        return cell
-    }
 }
